+++ routes/review.js
@@ -22,6 +22,10 @@ router.post(
   validateReview,
   wrapAsync(async (req, res) => {
     let listing = await Listing.findById(req.params.id);
+    if (!listing) {
+      req.flash("error", "Listing you requested for does not exist");
+      return res.redirect("/listings");
+    }
     let newReview = new Review(req.body.review);
 
     listing.reviews.push(newReview._id); //stores reference
@@ -37,6 +41,11 @@ router.delete(
   "/:reviewId",
   wrapAsync(async (req, res) => {
     let { id, reviewId } = req.params;
+    const listing = await Listing.findById(id);
+    if (!listing) {
+      req.flash("error", "Listing you requested for does not exist");
+      return res.redirect("/listings");
+    }
     await Listing.findByIdAndUpdate(id, { $pull: { reviews: reviewId } });
     await Review.findByIdAndDelete(reviewId);
